commit-stats: Cleanup formatting and ignore stable branches
repositories named <project>-<version.number> are omitted in the
assumption that these are tracking stable branches of projects.  This
may lose some commit #'s for stable maintenance work done on that
branch, but avoids doublebooking all the commits prior to the start of
the branch.

More clever would be to make the script count only the delta patches,
but I'm too lazy to do that.  Plus I suspect the count there is going to
be small anyway.